fix(shared): rename try[On]BeforeUnmount
